# src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.1.0",
    "title": "React lab API",
    "description": "API on Node.js for ITRex React.js Lab",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "contact": {
      "name": "Tsimafei Kinevich",
      "url": "https://itrexgroup.com/",
      "email": "tsimafei.kinevich@itrexgroup.com"
    }
  },
  "schemes": ["https"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
...
    "/api/resolutions": {
      "post": {
        "tags": ["Resolutions"],
        "summary": "Create new resolution",
        "security": [{
          "Bearer": []
        }],
        "parameters": [{
          "name": "Request body",
          "in": "body",
          "required": true,
          "schema": {
            "type": "object",
            "properties": {
              "resolution": {
                "example": "poisoning",
                "maxLength": 30,
                "minLength": 2,
                "type": "string"
              },
              "appointmentID": {
                "example": "b33c379e-3238-11ec-8d3d-0242ac130003",
                "type": "string"
              }
            }
          }
        }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/Resolution"
            }
          }
        }
      }
    },
    "/api/resolutions/{id}": {
      "parameters": [{
        "name": "id",
        "in": "path",
        "required": true,
        "description": "Resolution ID",
        "type": "string"
      }],
      "delete": {
        "tags": ["Resolutions"],
        "summary": "Delete a resolution",
        "security": [{
          "Bearer": []
        }],
        "responses": {
          "200": {
            "description": "Deleted",
            "schema": {
              "description": "Deleted resolution ID",
              "type": "string"
            }
          }
        }
      },
      "patch": {
        "tags": ["Resolutions"],
        "summary": "Update a resolution",
        "security": [{
          "Bearer": []
        }],
        "parameters": [
          {
            "name": "Request body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "resolution": {
                  "example": "poisoning",
                  "maxLength": 30,
                  "minLength": 2,
                  "type": "string"
                }
              }
            }
          }],
        "responses": {
          "200": {
            "description": "Updated",
            "schema": {
              "description": "Updated resolution ID",
              "type": "string"
            }
          }
        }
      }
    },
    "/api/resolutions/patient/me": {
      "get": {
        "tags": ["Resolutions"],
        "summary": "Get all resolutions for patient page",
        "security": [{
          "Bearer": []
        }],
        "parameters": [{
          "name": "offset",
          "in": "query",
          "description": "Offset for pagination",
          "required": true,
          "type": "number"
        },
          {
            "name": "limit",
            "in": "query",
            "description": "Limit for pagination",
            "required": true,
            "type": "number"
          },
          {
            "name": "name",
            "in": "query",
            "description": "Doctor name filter",
            "required": false,
            "type": "string"
          },
          {
            "name": "sortBy",
            "in": "query",
            "description": "Sort by",
            "required": false,
            "type": "string",
            "enum": ["firstNameSort", "lastNameSort", "dateSort", "nextDateSort"]
          },
          {
            "name": "order",
            "in": "query",
            "description": "Sort variant",
            "required": false,
            "type": "string",
            "enum": ["Asc", "Desc"]
          }
        ],
        "responses": {
          "200": {
            "description": "Resolutions",
            "schema": {
              "type": "object",
              "properties": {
                "resolutions": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/ResolutionForPatient"
                  }
                },
                "total": {
                  "type": "number"
                }
              }
            }
          }
        }
      }
    },
    "/api/resolutions/doctor/specialization/{id}": {
      "parameters": [{
        "name": "id",
        "in": "path",
        "required": true,
        "description": "Specialization ID",
        "type": "string"
      }],
      "get": {
        "tags": ["Resolutions"],
        "summary": "Filter by specializations for patient page resolutions",
        "security": [{
          "Bearer": []
        }],
        "parameters": [{
          "name": "offset",
          "in": "query",
          "description": "Offset for pagination",
          "required": true,
          "type": "number"
        },
          {
            "name": "limit",
            "in": "query",
            "description": "Limit for pagination",
            "required": true,
            "type": "number"
          },
          {
            "name": "name",
            "in": "query",
            "description": "Doctor name filter",
            "required": false,
            "type": "string"
          },
          {
            "name": "sortBy",
            "in": "query",
            "description": "Sort by",
            "required": false,
            "type": "string",
            "enum": ["firstNameSort", "lastNameSort", "dateSort", "nextDateSort"]
          },
          {
            "name": "order",
            "in": "query",
            "description": "Sort variant",
            "required": false,
            "type": "string",
            "enum": ["Asc", "Desc"]
          }
        ],
        "responses": {
          "200": {
            "description": "Resolutions",
            "schema": {
              "type": "object",
              "properties": {
                "resolutions": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/ResolutionForPatient"
                  }
                },
                "total": {
                  "type": "number"
                }
              }
            }
          }
        }
      }
    },
    "/api/resolutions/date": {
      "get": {
        "tags": ["Resolutions"],
        "summary": "Filter by dates for patient page resolutions",
        "security": [{
          "Bearer": []
        }],
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "description": "Date filter",
            "required": true,
            "type": "string"
          },
          {
          "name": "offset",
          "in": "query",
          "description": "Offset for pagination",
          "required": true,
          "type": "number"
        },
          {
            "name": "limit",
            "in": "query",
            "description": "Limit for pagination",
            "required": true,
            "type": "number"
          },
          {
            "name": "name",
            "in": "query",
            "description": "Doctor name filter",
            "required": false,
            "type": "string"
          },
          {
            "name": "sortBy",
            "in": "query",
            "description": "Sort by",
            "required": false,
            "type": "string",
            "enum": ["firstNameSort", "lastNameSort", "dateSort", "nextDateSort"]
          },
          {
            "name": "order",
            "in": "query",
            "description": "Sort variant",
            "required": false,
            "type": "string",
            "enum": ["Asc", "Desc"]
          }
        ],
        "responses": {
          "200": {
            "description": "Resolutions",
            "schema": {
              "type": "object",
              "properties": {
                "resolutions": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/ResolutionForPatient"
                  }
                },
                "total": {
                  "type": "number"
                }
              }
            }
          }
        }
      }
    },
    "/api/resolutions/doctor/me": {
      "get": {
        "tags": ["Resolutions"],
        "summary": "Get all resolutions for doctor page",
        "security": [{
          "Bearer": []
        }],
        "parameters": [{
          "name": "offset",
          "in": "query",
          "description": "Offset for pagination",
          "required": true,
          "type": "number"
        },
          {
            "name": "limit",
            "in": "query",
            "description": "Limit for pagination",
            "required": true,
            "type": "number"
          },
          {
            "name": "name",
            "in": "query",
            "description": "Doctor name filter",
            "required": false,
            "type": "string"
          },
          {
            "name": "sortBy",
            "in": "query",
            "description": "Sort by",
            "required": false,
            "type": "string",
            "enum": ["firstNameSort", "lastNameSort", "dateSort", "nextDateSort"]
          },
          {
            "name": "order",
            "in": "query",
            "description": "Sort variant",
            "required": false,
            "type": "string",
            "enum": ["Asc", "Desc"]
          }
        ],
        "responses": {
          "200": {
            "description": "Resolutions",
            "schema": {
              "type": "object",
              "properties": {
                "resolutions": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/ResolutionForDoctor"
                  }
                },
                "total": {
                  "type": "number"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Resolution": {
      "properties": {
        "id": {
          "type": "string"
        },
        "appointment_id": {
          "type": "string"
        },
        "next_appointment_date": {
          "type": "string"
        },
        "resolution": {
          "type": "string"
        }
      }
    },
    "ResolutionForPatient": {
      "properties": {
        "id": {
          "type": "string"
        },
        "appointment_id": {
          "type": "string"
        },
        "next_appointment_date": {
          "type": "string"
        },
        "resolution": {
          "type": "string"
        },
        "visit_date": {
          "type": "string"
        },
        "first_name": {
          "type": "string"
        },
        "last_name": {
          "type": "string"
        },
        "specialization_name": {
          "type": "string"
        }
      }
    },
    "ResolutionForDoctor": {
      "properties": {
        "id": {
          "type": "string"
        },
        "appointment_id": {
          "type": "string"
        },
        "next_appointment_date": {
          "type": "string"
        },
        "resolution": {
          "type": "string"
        },
        "visit_date": {
          "type": "string"
        },
        "first_name": {
          "type": "string"
        },
        "last_name": {
          "type": "string"
        }
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}